backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.5
httpx==0.27.2
python-multipart==0.0.9
orjson==3.10.7
motor==3.5.1
pymongo==4.8.0
bcrypt==4.1.2
PyJWT==2.8.0
email-validator==2.2.0
